requirements.txt
jupyterlab
numpy
uproot
pandas
matplotlib
